+++ Creepy/Tiles/ITile.cs
@@ -4,69 +4,6 @@ using Microsoft.Xna.Framework;
 namespace Creepy
 {
 	/// <summary>
-	/// Represents an object in the tile
-	/// </summary>
-	public interface ITileObject
-	{
-		/// <summary>
-		/// Gets the tile where this object belongs
-		/// </summary>
-		/// <value>The tile.</value>
-		ITile Tile { get; }
-	}
-
-	public class LavaGenerator : ITileObject
-	{
-		/// <summary>
-		/// Gets or sets the lava generation speed
-		/// </summary>
-		/// <value>The generation speed.</value>
-		public float GenerationSpeed { get; set; }
-
-		/// <summary>
-		/// Gets the tile where this object belongs
-		/// </summary>
-		public ITile Tile { get; }
-	}
-
-	public class TrainStationTileObject : ITileObject
-	{
-		/// <summary>
-		/// Gets the tile where this object belongs
-		/// </summary>
-		public ITile Tile { get; }
-
-		/// <summary>
-		/// Gets the train network
-		/// </summary>
-		/// <value>The network.</value>
-		public TrainNetwork Network { get; }
-	}
-
-	/// <summary>
-	/// Manager for a network of trains and stations
-	/// </summary>
-	public class TrainNetwork
-	{
-	}
-
-	/// <summary>
-	/// represents a tile object that can have population
-	/// </summary>
-	public class TownTileObject : ITileObject
-	{
-		/// <summary>
-		/// Gets the population of this <see cref="TownTileObject"/>
-		/// </summary>
-		public int Population { get; protected set; }
-
-		/// <summary>
-		/// Gets the tile where this object belongs
-		/// </summary>
-		public ITile Tile { get; }
-	}
-
-	/// <summary>
 	/// Represents a tile in the map
 	/// </summary>
 	public interface ITile
