+++ Creepy/Tiles/TrainNetwork.cs
@@ -1,0 +1,10 @@
+
+namespace Creepy
+{
+	/// <summary>
+	/// Manager for a network of trains and stations
+	/// </summary>
+	public class TrainNetwork
+	{
+	}
+}
